% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_P1_lhmm}
\alias{compute_P1_lhmm}
\title{Compute LHMM probabilities from parameters}
\usage{
compute_P1_lhmm(para_P1)
}
\arguments{
\item{para_P1}{a vector of length \code{K-1}. parameters of initial state probability vector}
}
\value{
initial state probability vector of length \code{K}
}
\description{
Compute initial state probability from LHMM parameters
}
\seealso{
\code{\link{compute_PQ_lhmm}} for state-transition and state-action
probabilities of LHMM, \code{\link{compute_paras_hmm}} for computing
probabilities in HMM.
}
