% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_state_seq}
\alias{find_state_seq}
\title{Viterbi algorithm for HMM}
\usage{
find_state_seq(seq, P1, P, Q)
}
\arguments{
\item{seq}{An action sequence coded in integers}

\item{P1}{initial state probability vector of length \code{K}}

\item{P}{\code{K} by \code{K} state transition probability matrix}

\item{Q}{\code{K} by \code{N} state-action (emission) probability matrix}
}
\value{
a hidden state sequence coded in integers
}
\description{
Find the most likely hidden state sequence of an observed sequence under HMM
}
