% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proclhmm-package.R
\docType{package}
\name{proclhmm-package}
\alias{proclhmm-package}
\title{proclhmm: Latent Hidden Markov Models for Response Process Data}
\description{
This package provides functions for simulating from and fitting the latent hidden Markov models for response process data (Tang, 2024) \doi{10.1007/s11336-023-09938-1}. It also includes functions for simulating from and fitting ordinary hidden Markov models.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/xytangtang/proclhmm/issues}
}

}
\author{
\strong{Maintainer}: Xueying Tang \email{xueyingtang1989@gmail.com} [copyright holder]

}
\keyword{internal}
