% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proclhmm-package.R
\docType{_PACKAGE}
\name{proclhmm}
\alias{proclhmm}
\title{proclhmm: Latent Hidden Markov Models for Response Process Data}
\description{
proclhmm: Latent Hidden Markov Models for Response Process Data
}
