% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_lhmm}
\alias{sim_lhmm}
\title{Simulating action sequences using LHMM}
\usage{
sim_lhmm(n, paras, min_length, mean_len, return_hidden_state = TRUE)
}
\arguments{
\item{n}{number of action sequences to be generated}

\item{paras}{a list containing specified LHMM parameters: \code{para_a}, \code{para_b}, \code{para_alpha},
\code{para_beta}, and \code{para_P1}.}

\item{mean_len}{mean length of generated sequences}

\item{return_hidden_state}{logical. Whether generated hidden state sequences should be returned or not.}

\item{min_len}{minimum length of generated sequences}
}
\value{
\code{sim_hmm} returns a list of \code{n} generated action sequences if \code{return_hidden_state = FALSE}.
If \code{return_hidden_state = TRUE}, it returns a list of two lists, \code{seqs} and \code{state_seqs}. \code{seqs} gives
the generated action sequences. \code{state_seqs} gives the corresponding hidden state sequences.
}
\description{
\code{sim_lhmm} generate \code{n} action sequences from LHMM based on given parameters.
The lengths of the generated sequences are simulated from a Poission distribution with
mean \code{mean_len} and at least \code{min_len}. The latent trait is generated from standard
normal.
}
