% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_lhmm_paras}
\alias{sim_lhmm_paras}
\title{generate LHMM parameters}
\usage{
sim_lhmm_paras(N, K)
}
\arguments{
\item{N}{number of distinct actions}

\item{K}{number of hidden states}
}
\value{
a list of five elements, \code{para_a}, \code{para_b}, \code{para_alpha}, \code{para_beta}, and \code{para_P1}.
}
\description{
\code{sim_hmm_paras} generates the parameters of LHMM with \code{K} hidden states
and \code{N} distinct actions from Uniform(-0.5, 0.5).
}
